Do not retry to migrate objects if there is a decryption error
If multiple encryption keys are used in a project, or if the path
encryption is different between buckets, a single access key is not
enough to migrate all items. To avoid reprocessing the same failed
objects over and over again, metasearch stores the queue time of the
last object that passed the decryption step, and queries objects that
have identical or greater queue time.